Web-Servers: Add angular-nginx workload sample
This is probably the most popular use case and doesn't have a
sample reference yet. There is some confusion in the community
with users trying the Nodejs buildpack along with NGINX buildpack
to build angular apps and complaining that it's not working.

It's basically the same app in the nodejs dir, but with nginx-based
build parameters.

I have moved the javascript-frontend-sample to react-frontend-sample

Nodejs samples:

Stick the following banner to all frontend apps to avoid any confusion
between the purpose of 2 buildpacks:

<i> Note: Node.js buildpack is designed to build backend applications that use Node as their server.
To build frontend applications and serve them with NGINX/HTTPD, please use the Web Servers buildpack </i>